% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_model.R
\name{brms_model}
\alias{brms_model}
\title{brms model wrapper}
\usage{
brms_model(
  y = NULL,
  samples = NULL,
  horizon = NULL,
  model = NULL,
  n_cores = 1,
  n_chains = 4,
  n_iter = 2000,
  ...
)
}
\arguments{
\item{y}{Numeric vector of time points to forecast}

\item{samples}{Numeric, number of samples to take.}

\item{horizon}{Numeric, the time horizon over which to predict.}

\item{model}{A \code{brms} model wrapped in the \code{bf} function}

\item{n_cores}{Numeric, the number of cores to use}

\item{n_chains}{Numeric, the number of chains to use}

\item{n_iter}{Numeric, the number of iterations in the sampler to use}

\item{...}{additional arguments passed to \code{brms} (e.g. priors or family)}
}
\value{
A dataframe of predictions (with columns representing the time horizon and rows representing samples).
}
\description{
brms model wrapper
}
\examples{

## Used on its own
brms_model(y = EpiSoon::example_obs_rts[1:10, ]$rt,
          model = bf(y ~ gp(time)),
          samples = 10, horizon = 7, ...)

## Used for forecasting
forecast_rt(EpiSoon::example_obs_rts[1:10, ],
            model = function(...){
              brms_model(model = bf(y ~ gp(time)), ...)},
                   horizon = 7, samples = 10)
}
