% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsts_model.R
\name{bsts_model}
\alias{bsts_model}
\title{BSTS model wrapper}
\usage{
bsts_model(y = NULL, samples = NULL, horizon = NULL, model = NULL)
}
\arguments{
\item{y}{Numeric vector of time points to forecast}

\item{samples}{Numeric, number of samples to take.}

\item{horizon}{Numeric, the time horizon over which to predict.}

\item{model}{A \code{bsts} model object wrapped in a function with an \code{ss} and \code{y} argument.}
}
\value{
A dataframe of predictions (with columns representing the time horizon and rows representing samples).
}
\description{
BSTS model wrapper
}
\examples{


forecast_rt(EpiSoon::example_obs_rts[1:10, ],
           model = function(...){bsts_model(model = function(ss, y){
           bsts::AddAutoAr(ss, y = y, lags = 10)}, ...)},
           horizon = 7, samples = 10)
}
