% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture_model_wrapper.R
\name{crps_ensemble}
\alias{crps_ensemble}
\title{Stack models according to CRPS}
\usage{
crps_ensemble(y = NULL, models = NULL, weighting_period = 5, ...)
}
\arguments{
\item{y}{Numeric vector of time points to forecast}

\item{models}{A list of models. Models must be analogous to the form
\code{function(...){EpiSoon::fable_model(model = , ...)}} or
\code{function(...){EpiSoon::bsts_model(model = , ...)}}.}
}
\value{
A dataframe of predictions (with columns representing the
time horizon and rows representing samples).
}
\description{
Provides a wrapper for different EpiSoon models and generates
a mixture model based on these models
}
\examples{

y <- EpiSoon::example_obs_rts[1:10, ]$rt
model = fable::ARIMA(y ~ time)
samples = 10
horizon = 7
weighting_period = 5

# make list with models
baseline_ensemble <- function(...) {
EpiSoon::fable_model(model = fabletools::combination_model(fable::ARIMA(y), fable::ETS(y), fable::NAIVE(y),
                                                           fable::RW(y ~ drift()), cmbn_args = list(weights = "inv_var")), ...)
}
models <- list(
  Mixfabl = baseline_ensemble,
  "ARIMA" = function(...){EpiSoon::fable_model(model = fable::ARIMA(y), ...)},
  "ETS" = function(...){EpiSoon::fable_model(model = fable::ETS(y), ...)},
  "Drift" = function(...){EpiSoon::fable_model(model = fable::RW(y ~ drift()), ...)}
)

# make forecst on its own
forecast <- crps_ensemble(y = y,
                          models = models,
                          samples = 10,
                          horizon = 7,
                          weighting_period = 5)


# together with forecast_rt
forecast_rt(EpiSoon::example_obs_rts[1:10, ],
           model = function(...){
             crps_ensemble(y = y,
                           models = models,
                           samples = 10,
                           horizon = 7,
                           weighting_period = 5)
                         })


}
