% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_from_si_prob.R
\name{draw_from_si_prob}
\alias{draw_from_si_prob}
\title{Draw from the Serial Interval Probability Distribution}
\usage{
draw_from_si_prob(days_ago = NULL, serial_interval = NULL)
}
\arguments{
\item{days_ago}{Numeric vector of days in the past. Defaults to \code{NULL}.}

\item{serial_interval}{A numeric vector describing the probability distribution the serial interval.
See \code{EpiNow::covid_serial_interval} for an example of the format.}
}
\value{
A draw from the probability distribution the serial interval.
}
\description{
Draw from the Serial Interval Probability Distribution
}
\examples{
## Example serial interval
mean_si <- 4.7
sd_si <- 2.9

## Example serial interval
mu_log <- log(mean_si) - 1/2 * log((sd_si / mean_si)^2 + 1)
sd_log <- sqrt(log((sd_si/mean_si)^2 + 1))


serial_interval <- rlnorm(1:100, mu_log, sd_log) \%>\%
   round(0) \%>\%
   table \%>\%
   {. / sum(.)}

## Draw
draw_from_si_prob(c(1, 4, 6), serial_interval)
}
