% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obs_cases.R
\docType{data}
\name{example_obs_cases}
\alias{example_obs_cases}
\title{Example Observed Cases}
\format{A data frame containing cases reported on each date.}
\usage{
example_obs_cases
}
\description{
An example data frame of observed cases
}
\keyword{datasets}
