% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_obs_rts.R
\docType{data}
\name{example_obs_rts}
\alias{example_obs_rts}
\title{Example Observed Rts}
\format{A data frame containing Rts estimated for each date.}
\usage{
example_obs_rts
}
\description{
An example data frame of observed Reproduction numbers
}
\keyword{datasets}
