% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_serial_interval.R
\docType{data}
\name{example_serial_interval}
\alias{example_serial_interval}
\title{Example Serial Interval}
\format{
A vector giviing the probability for each day
}
\usage{
example_serial_interval
}
\description{
An example serial interval probability vector
}
\keyword{datasets}
