% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_model.R
\name{fable_model}
\alias{fable_model}
\title{Fable model wrapper}
\usage{
fable_model(y = NULL, samples = NULL, horizon = NULL, model = NULL)
}
\arguments{
\item{model}{A \code{fable} model object wrapped in a function with an \code{ss} and \code{y} argument.}
}
\value{
A dataframe of predictions (with columns representing the time horizon and rows representing samples).
}
\description{
Fable model wrapper
}
\examples{
## Used on its own
fable_model(y = EpiSoon::example_obs_rts[1:10, ]$rt,
           model = fable::ARIMA(y ~ time),
           samples = 10, horizon = 7)


forecast_rt(EpiSoon::example_obs_rts[1:10, ],
           model = function(...){
           fable_model(model = fable::ARIMA(y ~ time), ...)},
           horizon = 7, samples = 10)
}
