% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit and Sample From a BSTS Model}
\usage{
fit_model(rts, model = model, horizon = 7, samples = 1000)
}
\arguments{
\item{rts}{A dataframe of containing two variables \code{rt} and  \code{date} with
\code{rt} being numeric and \code{date} being a date.}

\item{model}{An uninitialised \code{bsts} model}

\item{horizon}{Numeric, the time horizon over which to predict}

\item{samples}{Numeric, the number of samples to from the posterior of the model
fit. Note that twice this number of MCMC steps will be taken with half used as burn in.}
}
\value{
A dataframe of samples containing the following variables:
\code{sample}, \code{date}, \code{rt}, and \code{horizon}.
}
\description{
Fit and Sample From a BSTS Model
}
\examples{

rts <- data.frame(rt = 1:10,
                 date = as.Date("2020-01-01") + lubridate::days(1:10))


fit_model(rts, model = bsts::AddAutoAr, horizon = 7, samples = 1000)
}
