% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_cases_directly}
\alias{forecast_cases_directly}
\title{Forecasts Cases Directly}
\usage{
forecast_cases_directly(
  cases = NULL,
  model,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE,
  timeout = 100
)
}
\arguments{
\item{cases}{A dataframe containing \code{date} and \code{cases} variables}

\item{model}{A model object in the format of \code{bsts_model} or \code{fable_model}. See the corresponding
help files for details.}

\item{horizon}{Numeric, the time horizon over which to predict.}

\item{samples}{Numeric, number of samples to take.}

\item{bound_rt}{Logical, defaults to \code{TRUE}. Should Rt values be bounded to be greater than or
equal to 0.}

\item{timeout}{Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.}
}
\value{
Forecast cases over a future forecast horizon
}
\description{
Forecasts Cases Directly
}
\examples{
\dontrun{
forecast_cases_directly(EpiSoon::example_obs_cases,
                model = function(...){
                        EpiSoon::bsts_model(model = function(ss, y){
                        bsts::AddAutoAr(ss, y = y, lags = 10)}, ...)
                        },
                        horizon = 7, samples = 10)
}
}
