% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-wrappers.R
\name{forecast_model}
\alias{forecast_model}
\title{forecast model wrapper}
\usage{
forecast_model(y = NULL, samples = NULL, horizon = NULL, model = NULL)
}
\arguments{
\item{y}{Numeric vector of time points to forecast}

\item{samples}{Numeric, number of samples to take.}

\item{horizon}{Numeric, the time horizon over which to predict.}

\item{model}{A \code{forecast} model object.}
}
\value{
A dataframe of predictions (with columns representing the
time horizon and rows representing samples).
}
\description{
Allows users to forecast using models from the \code{forecast}
Note that \code{forecast} must be installed for this model wrapper to be functional.
}
\examples{
\dontrun{

## Used on its own
forecast_model(y = EpiSoon::example_obs_rts[1:10, ]$rt,
               model = forecast::auto.arima,
               samples = 10, horizon = 7)


forecast_rt(EpiSoon::example_obs_rts[1:10, ],
            model = function(...){
            forecast_model(model = forecast::ets, ...)},
            horizon = 7, samples = 10)
}

}
