% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_rt.R
\name{forecast_rt}
\alias{forecast_rt}
\title{Fit and Forecast using a BSTS Model}
\usage{
forecast_rt(
  rts,
  model = model,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE,
  timeout = 100
)
}
\arguments{
\item{rts}{A dataframe of containing two variables \code{rt} and  \code{date} with
\code{rt} being numeric and \code{date} being a date.}

\item{model}{A model object in the format of \code{bsts_model} or \code{fable_model}. See the corresponding
help files for details.}

\item{horizon}{Numeric, the time horizon over which to predict.}

\item{samples}{Numeric, number of samples to take.}

\item{bound_rt}{Logical, defaults to \code{TRUE}. Should Rt values be bounded to be greater than or
equal to 0.}

\item{timeout}{Numeric, timeout of model fitting in seconds. Defaults to 30 seconds.}
}
\value{
A dataframe of samples containing the following variables:
\code{sample}, \code{date}, \code{rt}, and \code{horizon}.
}
\description{
Fit and Forecast using a BSTS Model
}
\examples{
forecast_rt(EpiSoon::example_obs_rts[1:10, ],
            model = function(...){
            EpiSoon::bsts_model(model = function(ss, y){
            bsts::AddAutoAr(ss, y = y, lags = 10)}, ...)
            },
            horizon = 7, samples = 10)
}
