% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecastHybrid_model.R
\name{forecastHybrid_model}
\alias{forecastHybrid_model}
\title{BSTS model wrapper}
\usage{
forecastHybrid_model(y = NULL, samples = NULL, horizon = NULL, ...)
}
\arguments{
\item{y}{Numeric vector of time points to forecast}

\item{samples}{Numeric, number of samples to take.}

\item{horizon}{Numeric, the time horizon over which to predict.}
}
\value{
A dataframe of predictions (with columns representing the time horizon and rows representing samples).
}
\description{
BSTS model wrapper
}
\examples{
\dontrun{

library(forecastHybrid)

## Used on its own
forecastHybrid_model(y = EpiSoon::example_obs_rts$rt,
                     samples = 1, horizon = 7,
                     weights = "cv.errors", windowSize = 7, cvHorizon = 2) -> tmp

## Used for forecasting
 forecast_rt(EpiSoon::example_obs_rts,
                    model = EpiSoon::forecastHybrid_model,
                    horizon = 7, samples = 10)
}
}
