% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infectiousness.R
\name{infectiousness}
\alias{infectiousness}
\title{Calculate Infectiousness}
\usage{
infectiousness(
  cases,
  prediction_date = NULL,
  horizon = NULL,
  mean_si = NULL,
  sd_si = NULL
)
}
\arguments{
\item{cases}{A dataframe containing \code{date} and \code{cases} variables}

\item{prediction_date}{A character string date (format "yyyy-mm-dd") indicating
the prediction date.}

\item{horizon}{Numeric, the time horizon of the forecast.}

\item{mean_si}{Numeric, the mean of the serial interval}

\item{sd_si}{Numeric, the standard deviation of the serial interval}
}
\value{
The infectiousness at the current time
}
\description{
Calculate Infectiousness
}
\examples{

cases <- data.frame(date = seq(as.Date("2020-01-01"),
                               as.Date("2020-01-10"),
                               by = "days"),
                    cases = 1:10)

infectiousness(cases, prediction_date = "2020-01-10", horizon = 3,
               mean_si = 4.7, sd_si = 2.9)
}
