% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_model_fit.R
\name{iterative_model_fit}
\alias{iterative_model_fit}
\title{Iteratively fit a BSTS model to each timepoint}
\usage{
iterative_model_fit(
  rts,
  model = NULL,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE
)
}
\arguments{
\item{rts}{A dataframe of containing two variables \code{rt} and  \code{date} with
\code{rt} being numeric and \code{date} being a date.}

\item{model}{An uninitialised \code{bsts} model}

\item{horizon}{Numeric, the time horizon over which to predict}

\item{samples}{Numeric, the number of samples to from the posterior of the model
fit. Note that twice this number of MCMC steps will be taken with half used as burn in.}

\item{bound_rt}{Logical, defaults to \code{TRUE}. Should Rt values be bounded to be greater than or
equal to 0.}
}
\value{

}
\description{
Iteratively fit a BSTS model to each timepoint
}
\examples{

rts <- data.frame(rt = 1:10,
                 date = as.Date("2020-01-01") + lubridate::days(1:10))


iterative_model_fit(rts, model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                    horizon = 7, samples = 10)
}
