% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot a Forecast}
\usage{
plot_forecast(
  forecast = NULL,
  observation = NULL,
  horizon_cutoff = NULL,
  obs_cutoff_at_forecast = TRUE
)
}
\arguments{
\item{forecast}{A dataframe as produced by \code{fit_model}.}

\item{observation}{A dataframe of observations containing the following variables:
\code{rt} and \code{date}.}

\item{horizon_cutoff}{Numeric, defaults to NULL. Forecast horizon to plot up to.}

\item{obs_cutoff_at_forecast}{Logical defaults to \code{TRUE}. Should observations only be shown
up to the date of the forecast.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot a Forecast
}
\examples{


observations <- data.frame(rt = 1:20,
                           date = as.Date("2020-01-01") + lubridate::days(1:20))

## Fit a model
samples <- fit_model(observations[1:10, ],
                     model = bsts::AddAutoAr,
                     horizon = 7, samples = 10)

## Summarise forecast
summarised_forecast <- summarise_forecast(samples)

## Plot forecast
plot_forecast(summarised_forecast, observations)
}
