% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast_evaluation.R
\name{plot_forecast_evaluation}
\alias{plot_forecast_evaluation}
\title{Plot a Forecast}
\usage{
plot_forecast_evaluation(
  forecasts = NULL,
  observations = NULL,
  horizon_to_plot = 1
)
}
\arguments{
\item{forecasts}{A dataframe as produced by \code{forecast_rt}
or \code{forecast_cases}}

\item{observations}{A dataframe of observations containing the
following variables:
\itemize{
\item either \code{rt} or \code{cases}
\item and \code{date}.
}}

\item{horizon_to_plot}{Numeric vector, the forecast horizon to plot.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot a Forecast
}
\examples{


observations <- data.frame(rt = 1:20,
                           date = as.Date("2020-01-01")
                                  + lubridate::days(1:20))

## Evaluate a model
forecast_eval <- evaluate_model(observations,
                                model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                               horizon = 7, samples = 10)

forecasts <- forecast_eval$forecasts

## Plot forecast
plot_forecast_evaluation(forecasts, observations, horizon_to_plot = 7)
}
