% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_current_cases.R
\name{predict_current_cases}
\alias{predict_current_cases}
\title{Predict cases for Rts based on observed data}
\usage{
predict_current_cases(
  cases = NULL,
  rts = NULL,
  serial_interval = NULL,
  rdist = NULL
)
}
\arguments{
\item{cases}{A dataframe containing \code{date} and \code{cases} variables}

\item{rts}{A dataframe of containing two variables \code{rt} and  \code{date} with
\code{rt} being numeric and \code{date} being a date.}

\item{serial_interval}{A numeric vector describing the probability distribution the serial interval.
See \code{EpiNow::covid_serial_interval} for an example of the format.}

\item{rdist}{A function to be used to sample the number of cases. Must take two
arguments with the first specfying the number of samples and the second the mean. Defaults
to \code{rpois} if not supplied}
}
\value{
Forecast cases for the current timestep
}
\description{
Predict cases for Rts based on observed data
}
\examples{


predict_current_cases(cases = EpiSoon::example_obs_cases,
                      rts = EpiSoon::example_obs_rts,
                      serial_interval = EpiSoon::example_serial_interval)
}
