% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_incidences.R
\name{predict_incidences}
\alias{predict_incidences}
\title{Predict incidence from Rts}
\usage{
predict_incidences(rts = NULL, days_ahead = NULL)
}
\arguments{
\item{rts}{A dataframe containing two variables: \code{rt} and \code{date}}

\item{days_ahead}{Numeric the number of days ahead}
}
\value{

}
\description{
Predict incidence from Rts
}
\examples{

## Code
predict_incidences
}
