% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_case_forecast.R
\name{score_case_forecast}
\alias{score_case_forecast}
\title{Score a case forecast}
\usage{
score_case_forecast(pred_cases, obs_cases)
}
\arguments{
\item{pred_cases}{Dataframe of predicted cases with the following variables: \code{sample}, \code{date},
\code{cases} and forecast horizon. As produced by \code{forecast_cases}.}

\item{obs_cases}{Dataframe of observed cases with the following variables: \code{date} and \code{cases}.}
}
\value{
A dataframe containing the following scores per forecast timepoint: dss, crps,
logs, bias, and sharpness as well as the forecast date and time horizon.
}
\description{
Score a case forecast
}
\examples{

## Observed cases
cases <- data.frame(date = seq(as.Date("2020-01-01"),
                               as.Date("2020-01-20"),
                               by = "days"),
                    cases = 1:20)
## Observed rts
observations <- data.frame(rt = 1:20,
                           date = as.Date("2020-01-01")
                                  + lubridate::days(1:20))

## Fit a model (using a subset of observations)
samples <- forecast_rt(observations[1:10, ],
                     model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                     horizon = 7, samples = 10)
## Example serial interval
mean_si <- 4.7
sd_si <- 2.9

mu_log <- log(mean_si) - 1/2 * log((sd_si / mean_si)^2 + 1)
sd_log <- sqrt(log((sd_si/mean_si)^2 + 1))


serial_interval <- rlnorm(1:100, mu_log, sd_log) \%>\%
   round(0) \%>\%
   table \%>\%
   {. / sum(.)}

pred_cases <- forecast_cases(cases[1:10, ], samples, serial_interval)

## Score the model fit (with observations during the time horizon of the forecast)
score_case_forecast(pred_cases, cases)
}
