% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_model.R
\name{score_model}
\alias{score_model}
\title{Score a Model Fit}
\usage{
score_model(fit_samples, observations)
}
\arguments{
\item{fit_samples}{A dataframe as produced by \code{EpiSoon::fit_model}.}

\item{observations}{A dataframe of observations against which to score. Should contain a \code{date} and \code{rt} column.}
}
\value{
A dataframe containing the following scores per forecast timepoint: dss, crps,
logs, bias, and sharpness as well as the forecast date and time horizon.
}
\description{
Score a Model Fit
}
\examples{

## Observed data
observations <- data.frame(rt = 1:20,
                           date = as.Date("2020-01-01")
                                  + lubridate::days(1:20))

## Fit a model
samples <- fit_model(observations[1:10, ],
                     model = bsts::AddAutoAr,
                     horizon = 7, samples = 10)

## Score the model fit
score_model(samples, observations)
}
