% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_case_forecast.R
\name{summarise_case_forecast}
\alias{summarise_case_forecast}
\title{Summarise Forecast Cases}
\usage{
summarise_case_forecast(pred_cases)
}
\arguments{
\item{fit_samples}{A dataframe as produced by \code{EpiSoon::forecast_cases}.}
}
\value{
A summarised dataframe.
}
\description{
Summarise Forecast Cases
}
\examples{

## Observed cases
cases <- data.frame(date = seq(as.Date("2020-01-01"),
                               as.Date("2020-01-10"),
                               by = "days"),
                    cases = 1:10)

## Forecast Rts
rts <- data.frame(date = seq(as.Date("2020-01-01"),
                               as.Date("2020-01-10"),
                               by = "days"),
                   rt = rnorm(10, 1.2, 0.01))


forecast <- forecast_rt(rts,
                        model = function(ss, y) {
                        bsts::AddAutoAr(ss, y = y, lags = 7)
                        },
                        horizon = 7, samples = 10)


case_forecast <- forecast_cases(cases, forecast, EpiSoon::example_serial_interval)

summarise_case_forecast(case_forecast)
}
