% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_fit.R
\name{summarise_fit}
\alias{summarise_fit}
\title{Summarise Predicted Rts}
\usage{
summarise_fit(fit_samples)
}
\arguments{
\item{fit_samples}{A dataframe as produced by \code{EpiSoon::fit_model}.}
}
\value{
A summarised dataframe.
}
\description{
Summarise Predicted Rts
}
\examples{

rts <- data.frame(rt = 1:10,
                 date = as.Date("2020-01-01") + lubridate::days(1:10))


samples <- fit_model(rts, model = bsts::AddAutoAr, horizon = 7, samples = 10)


summarise_fit(samples)
}
