% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_scores.R
\name{summarise_scores}
\alias{summarise_scores}
\title{Summarise model forecasting scores}
\usage{
summarise_scores(scores, variables = NULL)
}
\arguments{
\item{scores}{A dataframe of model scores as produced by \code{score_model}}

\item{variables}{A character vector of variables names to group over. By default score type
and model is grouped over if present.}
}
\value{
A dataframe of summarised scores in a tidy format.
}
\description{
Summarise model forecasting scores
}
\examples{

## Dummy data
observations <- data.frame(rt = 1:20,
                            date = as.Date("2020-01-01")
                           + lubridate::days(1:20))

timeseries <- list(observations, observations)
names(timeseries) <- c("Region 1", "Region 2")

## List of forecasting bsts models wrapped in functions.
models <- list("Sparse AR" = function(ss, y){bsts::AddAutoAr(ss, y = y, lags = 7)},
               "Semi-local linear trend" = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)})



## Compare models
evaluations <- compare_timeseries(timeseries, models,
                                  horizon = 7, samples = 10)


scores <- evaluations$scores


## Score across the default groups
summarise_scores(scores)


## Also summarise across time horizon
summarise_scores(scores, "horizon")

## Instead summarise across region
summarise_scores(scores, "region")
}
