% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_cases_x_days_ago.R
\name{weight_cases_x_days_ago}
\alias{weight_cases_x_days_ago}
\title{Weight cases x days ago}
\usage{
weight_cases_x_days_ago(num_days_ago = NULL, mean_si = NULL, sd_si = NULL)
}
\arguments{
\item{num_days_ago}{Numeric of days in the past. Defaults to \code{NULL}.}

\item{mean_si}{Numeric, the mean of the serial interval}

\item{sd_si}{Numeric, the standard deviation of the serial interval}
}
\value{
A weighting for cases x days ago (numeric)
}
\description{
Weight cases x days ago
}
\examples{


weight_cases_x_days_ago(10, 4.7, 2.9)
}
