% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brms_model.R
\name{brms_model}
\alias{brms_model}
\title{brms model wrapper}
\usage{
brms_model(
  y = NULL,
  samples = NULL,
  horizon = NULL,
  model = NULL,
  n_cores = 1,
  n_chains = 4,
  n_iter = 2000,
  ...
)
}
\arguments{
\item{y}{Numeric vector of time points to forecast}

\item{samples}{Numeric, number of samples to take.}

\item{horizon}{Numeric, the time horizon over which to predict.}

\item{model}{A \link{brms} model wrapped in the \code{\link[brms:bf]{brms::bf()}} function}

\item{n_cores}{Numeric, the number of cores to use, default of 1}

\item{n_chains}{Numeric, the number of chains to use, default of 4}

\item{n_iter}{Numeric, the number of iterations in the sampler to use,
default of 4000}

\item{...}{additional arguments passed to \code{brms} (e.g. priors or family)}
}
\value{
A dataframe of predictions (with columns representing the time horizon and rows representing samples).
}
\description{
Allows users to specify a model using the \code{\link[brms:bf]{brms::bf()}}  wrapper from \code{brms}
Note that \code{brms} and \code{tidybayes} must both be installed for this
model wrapper to be functional.
}
\examples{

## Used on its own
## Note: More iterations and chains should be used
library(brms)
brms_model(y = EpiSoon::example_obs_rts[1:10, ]$rt,
          model = brms::bf(y ~ gp(time)),
          samples = 10, horizon = 7, n_iter = 40, n_chains = 1, refresh =0)

## Used for forecasting
## Note that the timeout parameter has been increased to allow
## for the time for the code to be compiled
## Note: More iterations and chains should be used

forecast_rt(EpiSoon::example_obs_rts[1:10, ],
            model = function(...){
              brms_model(model = brms::bf(y ~ gp(time)), n_iter = 40, n_chains = 1, ...)},
                   horizon = 7, samples = 10, timeout = 300)
}
