% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_timeseries.R
\name{compare_timeseries}
\alias{compare_timeseries}
\title{Compare timeseries and forecast models}
\usage{
compare_timeseries(
  obs_rts = NULL,
  obs_cases = NULL,
  models = NULL,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE,
  min_pounts = 3,
  timeout = 30,
  serial_interval = NULL,
  rdist = NULL
)
}
\arguments{
\item{obs_rts}{A dataframe of observed Rts including a \code{timeseries} variable to denote
each timeseris and a  \code{rt} vector (to forecast) and a \code{date} vector (to denote time). Optionally this dataframe can
contain samples for each timeseries in which case this should be denoted using a \code{sample} variable.}

\item{obs_cases}{A dataframe of observed cases including a \code{timeseries} variable to denote
each timeseris and a  \code{cases} vector (to forecast) and a \code{date} vector (to denote time). Optionally this dataframe can
contain samples for each timeseries in which case this should be denoted using a \code{sample} variable.}

\item{models}{A list of \code{bsts} models. An example configuration is given in the
examples. Each \code{bsts} model needs to be wrapped in a function that takes a \code{ss} and \code{y}
argument (i.e. \code{function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}}).}

\item{horizon}{Numeric, the time horizon over which to predict}

\item{samples}{Numeric, the number of samples to from the posterior of the model
fit. Note that twice this number of MCMC steps will be taken with half used as burn in.}

\item{bound_rt}{Logical, defaults to \code{TRUE}. Should Rt values be bounded to be greater than or
equal to 0.}

\item{timeout}{Numeric, the number of seconds to allow before terminating model fitting. Defaults to
30 seconds.}
}
\value{
A list of dataframes as produced by \verb{evaluate model} but with an additional model column.
}
\description{
Compare timeseries and forecast models
}
\examples{

## Example data
obs_rts <- EpiSoon::example_obs_rts \%>\%
    dplyr::mutate(timeseries = "Region 1") \%>\%
    dplyr::bind_rows(EpiSoon::example_obs_rts \%>\%
    dplyr::mutate(timeseries = "Region 2"))

obs_cases <- EpiSoon::example_obs_cases \%>\%
    dplyr::mutate(timeseries = "Region 1") \%>\%
    dplyr::bind_rows(EpiSoon::example_obs_cases \%>\%
    dplyr::mutate(timeseries = "Region 2"))

## List of forecasting bsts models wrapped in functions.
models <- list("AR 3" =
                    function(ss, y){bsts::AddAr(ss, y = y, lags = 3)},
               "Semi-local linear trend" =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)})


## Compare models
evaluations <- compare_timeseries(obs_rts, obs_cases, models,
                                  horizon = 7, samples = 10,
                                  serial_interval = EpiSoon::example_serial_interval)

evaluations

## Example evaluation plot for comparing forecasts
## with actuals for a range of models and timeseries.
plot_forecast_evaluation(evaluations$forecast_rts, obs_rts, c(7)) +
  ggplot2::facet_grid(model ~ timeseries) +
  cowplot::panel_border()

## Hack to plot observed cases vs predicted
plot_forecast_evaluation(evaluations$forecast_cases,
                         obs_cases, c(7)) +
  ggplot2::facet_grid(model ~ timeseries, scales = "free") +
  cowplot::panel_border()
}
