% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{draw_from_si_prob}
\alias{draw_from_si_prob}
\title{Draw from the Serial Interval Probability Distribution}
\usage{
draw_from_si_prob(days_ago = NULL, serial_interval = NULL)
}
\arguments{
\item{days_ago}{Numeric vector of days in the past. Defaults to \code{NULL}.}

\item{serial_interval}{A numeric vector describing the probability distribution the serial interval.
See \code{EpiNow::covid_serial_interval} for an example of the format.}
}
\value{
A draw from the probability distribution the serial interval.
}
\description{
Draw from the Serial Interval Probability Distribution
}
\examples{

## Draw
draw_from_si_prob(rev(c(1, 2, 4, 10, 1:100)), EpiSoon::example_serial_interval)
}
