% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate a Model for Forecasting Rts}
\usage{
evaluate_model(
  observations = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE
)
}
\arguments{
\item{observations}{A dataframe of observations against which to score. Should contain a \code{date} and \code{rt} column.}

\item{model}{An uninitialised \code{bsts} model}

\item{horizon}{Numeric, the time horizon over which to predict}

\item{samples}{Numeric, the number of samples to from the posterior of the model
fit. Note that twice this number of MCMC steps will be taken with half used as burn in.}

\item{bound_rt}{Logical, defaults to \code{TRUE}. Should Rt values be bounded to be greater than or
equal to 0.}
}
\value{

}
\description{
Evaluate a Model for Forecasting Rts
}
\examples{
## Observed data
observations <- data.frame(rt = 1:20,
                           date = as.Date("2020-01-01")
                                  + lubridate::days(1:20))

## Evaluate a model
evaluate_model(observations,
               model = bsts::AddAutoAr,
               horizon = 7, samples = 10)


}
