% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_model.R
\name{evaluate_model}
\alias{evaluate_model}
\title{Evaluate a Model for Forecasting Rts}
\usage{
evaluate_model(
  observations = NULL,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE
)
}
\arguments{
\item{observations}{A dataframe of observations to forecast with and score
against. Should contain a \code{date} and \code{rt} column. If multiple samples are included this
should be denoted using a numeric \code{sample} variable.}

\item{model}{An uninitialised \code{bsts} model}

\item{horizon}{Numeric, the time horizon over which to predict}

\item{samples}{Numeric, the number of samples to from the posterior of the model
fit. Note that twice this number of MCMC steps will be taken with half used as burn in.}

\item{timeout}{Numeric, the number of seconds to allow before terminating model fitting. Defaults to
30 seconds.}

\item{bound_rt}{Logical, defaults to \code{TRUE}. Should Rt values be bounded to be greater than or
equal to 0.}
}
\value{

}
\description{
Evaluate a Model for Forecasting Rts
}
\examples{
## Observed data
observations <- data.frame(rt = 1:20,
                           date = as.Date("2020-01-01")
                                  + lubridate::days(1:20))

## Evaluate a model
evaluate_model(observations,
               model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
               horizon = 7, samples = 10)


## Samples of observed data
sampled_obs <- observations \%>\%
   dplyr::mutate(sample = 1) \%>\%
   dplyr::bind_rows(observations \%>\%
       dplyr::mutate(sample = 2))

## Evaluate a model across samples
evaluate_model(sampled_obs,
               model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
               horizon = 7, samples = 10)
}
