% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fable_model.R
\name{fable_model}
\alias{fable_model}
\title{Fable model wrapper}
\usage{
fable_model(y = NULL, samples = NULL, horizon = NULL, model = NULL)
}
\arguments{
\item{y}{Numeric vector of time points to forecast}

\item{samples}{Numeric, number of samples to take.}

\item{horizon}{Numeric, the time horizon over which to predict.}

\item{model}{A \code{fable} model object. For  models that use a formula interface time
can be accessed using \code{time}.}
}
\value{
A dataframe of predictions (with columns representing the
time horizon and rows representing samples).
}
\description{
Provides an interface for models from the \code{fable} package.
Note the \code{feasts::ARIMA} model requires the \code{feast} package. If \code{future}
is being used \code{fable} will require \code{future.apply} in
order to not silently fail.
}
\examples{
\dontrun{
## Used on its own
fable_model(y = EpiSoon::example_obs_rts[1:10, ]$rt,
           model = fable::ARIMA(y ~ time),
           samples = 10, horizon = 7)


forecast_rt(EpiSoon::example_obs_rts[1:10, ],
           model = function(...){
           fable_model(model = fable::ARIMA(y ~ time), ...)},
           horizon = 7, samples = 10)
}
}
