% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_cases.R
\name{forecast_cases}
\alias{forecast_cases}
\title{Forecasts cases for a Rt forecasts}
\usage{
forecast_cases(
  cases = NULL,
  fit_samples = NULL,
  serial_interval = NULL,
  forecast_date = NULL,
  horizon = NULL,
  rdist = NULL
)
}
\arguments{
\item{cases}{A dataframe containing \code{date} and \code{cases} variables}

\item{fit_samples}{A dataframe as produced by \code{EpiSoon::forecast}.}

\item{serial_interval}{A numeric vector describing the probability distribution the serial interval.
See \code{EpiNow::covid_serial_interval} for an example of the format.}

\item{forecast_date}{A character string date (format "yyyy-mm-dd") indicating
the forecast date. Defaults to \code{NULL} in which case it will be assumed that the
forecast date is the  day before the first date present in the \code{fit_samples} dataframe}

\item{horizon}{Numeric, the time horizon over which to predict.}

\item{rdist}{A function to be used to sample the number of cases. Must take two
arguments with the first specfying the number of samples and the second the mean. Defaults
to \code{rpois} if not supplied}
}
\value{
Forecast cases for over a future forecast horizon
}
\description{
Forecasts cases for a Rt forecasts
}
\examples{
## Rt forecast
forecast <- forecast_rt(EpiSoon::example_obs_rts[1:10, ],
                        model = function(...){
                        EpiSoon::bsts_model(model = function(ss, y){
                        bsts::AddAutoAr(ss, y = y, lags = 10)}, ...)
                        },
                        horizon = 7, samples = 10)


forecast_cases(EpiSoon::example_obs_cases,
               fit_samples =  forecast,
               serial_interval = EpiSoon::example_serial_interval)
}
