% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast_rt.R
\name{forecast_rt}
\alias{forecast_rt}
\title{Fit and Forecast using a BSTS Model}
\usage{
forecast_rt(
  rts,
  model = model,
  horizon = 7,
  samples = 1000,
  bound_rt = TRUE,
  timeout = 30
)
}
\arguments{
\item{rts}{A dataframe of containing two variables \code{rt} and  \code{date} with
\code{rt} being numeric and \code{date} being a date.}

\item{model}{An uninitialised \code{bsts} model}

\item{horizon}{Numeric, the time horizon over which to predict}

\item{samples}{Numeric, the number of samples to from the posterior of the model
fit. Note that twice this number of MCMC steps will be taken with half used as burn in.}

\item{bound_rt}{Logical, defaults to \code{TRUE}. Should Rt values be bounded to be greater than or
equal to 0.}

\item{timeout}{Numeric, the number of seconds to allow before terminating model fitting. Defaults to
30 seconds.}
}
\value{
A dataframe of samples containing the following variables:
\code{sample}, \code{date}, \code{rt}, and \code{horizon}.
}
\description{
Fit and Forecast using a BSTS Model
}
\examples{
forecast_rt(EpiSoon::example_obs_rts[1:10, ],
            model = function(ss, y){bsts::AddAutoAr(ss, y = y, lags = 10)},
            horizon = 7, samples = 10)
}
