% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-wrappers.R
\name{forecastHybrid_model}
\alias{forecastHybrid_model}
\title{forecastHybrid Model Wrapper}
\usage{
forecastHybrid_model(
  y = NULL,
  samples = NULL,
  horizon = NULL,
  model_params = NULL,
  forecast_params = NULL
)
}
\arguments{
\item{y}{Numeric vector of time points to forecast}

\item{samples}{Numeric, number of samples to take.}

\item{horizon}{Numeric, the time horizon over which to predict.}

\item{model_params}{List of parameters to pass to \code{forecastHybrid::hybridModel}.}

\item{forecast_params}{List of parameters to pass to \code{forecastHybrid:::forecast.hybridModel}.}
}
\value{
A dataframe of predictions (with columns representing the time horizon and rows representing samples).
}
\description{
Allows users to forecast using ensembles from the \code{forecastHybrid} package. Note that
whilst weighted ensembles can be created this is not advised when samples > 1 as currently
samples are derived assuming a normal distribution using the upper and lower confidence intervals of the ensemble.
These confidence intervals are themselves either based on the unweighted mean of the ensembled
models or the maximum/minimum from the candiate models. Note that \code{forecastHybrid} must be installed for this
model wrapper to be functional.
}
\examples{
\dontrun{

library(forecastHybrid)

## Used on its own
forecastHybrid_model(y = EpiSoon::example_obs_rts$rt,
                     samples = 10, horizon = 7)


## Used with non-default arguments
## Note that with the current sampling from maximal confidence intervals model
## Weighting using cross-validation will only have an impact when 1 sample is used.
forecastHybrid_model(y = EpiSoon::example_obs_rts$rt,
                    samples = 1, horizon = 7,
                    model_params = list(cvHorizon = 7, windowSize = 7,
                                  rolling = TRUE, models = "zeta"))


## Used for forecasting
 forecast_rt(EpiSoon::example_obs_rts,
                    model = EpiSoon::forecastHybrid_model,
                    horizon = 7, samples = 1)

## Used for forcasting with non-default arguments
forecast_rt(EpiSoon::example_obs_rts,
           model = function(...){EpiSoon::forecastHybrid_model(
           model_params = list(models = "zte"),
           forecast_params = list(PI.combination = "mean"), ...)
           },
           horizon = 7, samples = 10)
}
}
