% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_rt_forecast.R
\name{iterative_rt_forecast}
\alias{iterative_rt_forecast}
\title{Iteratively forecast using a BSTS model}
\usage{
iterative_rt_forecast(
  rts,
  model = NULL,
  horizon = 7,
  samples = 1000,
  timeout = 30,
  bound_rt = TRUE,
  min_points = 3
)
}
\arguments{
\item{rts}{A dataframe of containing two variables \code{rt} and  \code{date} with
\code{rt} being numeric and \code{date} being a date.}

\item{model}{An uninitialised \code{bsts} model}

\item{horizon}{Numeric, the time horizon over which to predict}

\item{samples}{Numeric, the number of samples to from the posterior of the model
fit. Note that twice this number of MCMC steps will be taken with half used as burn in.}

\item{timeout}{Numeric, the number of seconds to allow before terminating model fitting. Defaults to
30 seconds.}

\item{bound_rt}{Logical, defaults to \code{TRUE}. Should Rt values be bounded to be greater than or
equal to 0.}

\item{min_points}{Numeric, defaults to 3. The minimum number of time points at which to begin
iteratively evaluating the forecast.}
}
\value{

}
\description{
Iteratively forecast using a BSTS model
}
\examples{


iterative_rt_forecast(EpiSoon::example_obs_rts,
                      model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                      horizon = 7, samples = 10, min_points = 4)
}
