% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_forecast.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot a Forecast}
\usage{
plot_forecast(
  forecast = NULL,
  observations = NULL,
  horizon_cutoff = NULL,
  obs_cutoff_at_forecast = TRUE
)
}
\arguments{
\item{forecast}{A dataframe with summarised forecasts as produced
by \code{summarise_forecast} or \code{summarise_case_forecast} .}

\item{observations}{A dataframe of observations containing the
following variables:
\itemize{
\item either \code{rt} or \code{cases}
\item and \code{date}.
}}

\item{horizon_cutoff}{Numeric, defaults to NULL. Forecast horizon to plot up to.}

\item{obs_cutoff_at_forecast}{Logical defaults to \code{TRUE}. Should observations only be shown
up to the date of the forecast.}
}
\value{
A \code{ggplot2} object
}
\description{
Plot a Forecast
}
\examples{

## Forecast an Rt sample
samples <- forecast_rt(EpiSoon::example_obs_rts[1:10, ],
                     model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                     horizon = 7, samples = 10)

## Summarise forecast
summarised_forecast <- summarise_forecast(samples)

## Plot forecast
plot_forecast(summarised_forecast, observations)
}
