% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_cases.R
\name{predict_cases}
\alias{predict_cases}
\title{Predict cases for a single Rt sample forecasts}
\usage{
predict_cases(
  cases = NULL,
  rts = NULL,
  serial_interval = NULL,
  forecast_date = NULL,
  horizon = NULL,
  rdist = NULL
)
}
\arguments{
\item{cases}{A dataframe containing \code{date} and \code{cases} variables}

\item{rts}{A dataframe of containing two variables \code{rt} and  \code{date} with
\code{rt} being numeric and \code{date} being a date.}

\item{serial_interval}{A numeric vector describing the probability distribution the serial interval.
See \code{EpiNow::covid_serial_interval} for an example of the format.}

\item{forecast_date}{A character string date (format "yyyy-mm-dd") indicating
the forecast date. Defaults to \code{NULL} in which case it will be assumed that the
forecast date is the last data present in the \code{cases} dataframe}

\item{horizon}{Numeric, the time horizon over which to predict.}

\item{rdist}{A function to be used to sample the number of cases. Must take two
arguments with the first specfying the number of samples and the second the mean. Defaults
to \code{rpois} if not supplied}
}
\value{
Forecast cases for over a future forecast horizon
}
\description{
Predict cases for a single Rt sample forecasts
}
\examples{

forecast <- forecast_rt(EpiSoon::example_obs_rts[1:12, ],
                        model = function(...) {EpiSoon::bsts_model(model =
                                function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
                        horizon = 7, samples = 1)


purrr::map_dfr(1:100, ~ predict_cases(cases = EpiSoon::example_obs_cases,
              rts = forecast,
              forecast_date = as.Date("2020-03-10"),
              serial_interval = example_serial_interval)) \%>\%
              dplyr::group_by(date) \%>\%
              dplyr::summarise(cases = mean(cases))
}
