% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_case_forecast.R
\name{score_case_forecast}
\alias{score_case_forecast}
\title{Score a Model Fit}
\usage{
score_case_forecast(pred_cases, obs_cases)
}
\arguments{
\item{observations}{A dataframe of observations against which to score. Should contain a \code{date} and \code{rt} column.}
}
\value{
A dataframe containing the following scores per forecast timepoint: dss, crps,
logs, bias, and sharpness as well as the forecast date and time horizon.
}
\description{
Score a Model Fit
}
\examples{

## Observed data
observations <- data.frame(rt = 1:20,
                           date = as.Date("2020-01-01")
                                  + lubridate::days(1:20))

## Fit a model (using a subset of observations)
samples <- forecast_rt(observations[1:10, ],
                     model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                     horizon = 7, samples = 10)
pred_cases <- predict_cases()

## Score the model fit (with observations during the time horizon of the forecast)
score_forecast(samples, observations)
}
