% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_case_forecast.R
\name{score_case_forecast}
\alias{score_case_forecast}
\title{Score a case forecast}
\usage{
score_case_forecast(pred_cases, obs_cases)
}
\arguments{
\item{pred_cases}{Dataframe of predicted cases with the following variables: \code{sample}, \code{date},
\code{cases} and forecast horizon. As produced by \code{forecast_cases}.}

\item{obs_cases}{Dataframe of observed cases with the following variables: \code{date} and \code{cases}.}
}
\value{
A dataframe containing the following scores per forecast timepoint: dss, crps,
logs, bias, and sharpness as well as the forecast date and time horizon.
}
\description{
Score a case forecast
}
\examples{
## Fit a model (using a subset of observations)
samples <- forecast_rt(EpiSoon::example_obs_rts[1:10, ],
                     model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                     horizon = 7, samples = 10)

pred_cases <- forecast_cases(EpiSoon::example_obs_cases[1:10, ],
                             samples, EpiSoon::example_serial_interval)

## Score the model fit (with observations during the time horizon of the forecast)
score_case_forecast(pred_cases, EpiSoon::example_obs_cases)
}
