% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_case_forecast}
\alias{summarise_case_forecast}
\title{Summarise Forecast Cases}
\usage{
summarise_case_forecast(pred_cases)
}
\arguments{
\item{pred_cases}{A dataframe as produced by \code{EpiSoon::forecast_cases}.}
}
\value{
A summarised dataframe.
}
\description{
Summarise Forecast Cases
}
\examples{
\dontrun{
## Example forecast
forecast <- forecast_rt(EpiSoon::example_obs_rts,
  model = function(...) {
    EpiSoon::bsts_model(
      model =
        function(ss, y) {
          bsts::AddSemilocalLinearTrend(ss, y = y)
        }, ...
    )
  },
  horizon = 7, samples = 10
)

## Forecast cases
case_forecast <- forecast_cases(
  EpiSoon::example_obs_cases,
  forecast,
  EpiSoon::example_serial_interval
)
## Summarise case forecast
summarise_case_forecast(case_forecast)
}
}
