% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_forecast.R
\name{summarise_forecast}
\alias{summarise_forecast}
\title{Summarise Forecast Rts}
\usage{
summarise_forecast(fit_samples)
}
\arguments{
\item{fit_samples}{A dataframe as produced by \code{EpiSoon::fit_model}.}
}
\value{
A summarised dataframe.
}
\description{
Summarise Forecast Rts
}
\examples{

rts <- data.frame(rt = 1:10,
                 date = as.Date("2020-01-01") + lubridate::days(1:10))


samples <- forecast_rt(rts, model = function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)},
                       horizon = 7, samples = 10)


summarise_forecast(samples)
}
