% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise_forecast}
\alias{summarise_forecast}
\title{Summarise Forecast Rts}
\usage{
summarise_forecast(fit_samples)
}
\arguments{
\item{fit_samples}{A dataframe as produced by \code{EpiSoon::forecast}.}
}
\value{
A summarised dataframe.
}
\description{
Summarise Forecast Rts
}
\examples{


samples <- forecast_rt(example_obs_rts,
                       model = function(...) {EpiSoon::bsts_model(model =
                    function(ss, y){bsts::AddSemilocalLinearTrend(ss, y = y)}, ...)},
                       horizon = 7, samples = 10)


summarise_forecast(samples)
}
